move custom calendar methods to base calendar
