+++ s2spy/_base_calendar.py
@@ -6,6 +6,8 @@ customized for each specific calendar.
 """
 from abc import ABC
 from abc import abstractmethod
+import re
+from typing import Tuple
 from typing import Union
 import numpy as np
 import pandas as pd
@@ -27,18 +29,18 @@ class BaseCalendar(ABC):
     def __init__(
         self,
         anchor,
-        freq,
-        n_targets: int = 1,
     ):
         """For initializing calendars, the following five variables will be required."""
-        self.n_targets = n_targets
-        self.anchor = anchor
-        self.freq = freq
+        self._anchor, self._anchor_fmt = self._parse_anchor(anchor)
+        self._targets: list[TargetPeriod] = []
+        self._precursors: list[PrecursorPeriod] = []
+        self._total_length_target = 0
+        self._total_length_precursor = 0
+        self.n_targets = 0
 
-        self._max_lag = 0
-        self._allow_overlap = False
+        self._allow_overlap: bool = False
 
-    @abstractmethod
+
     def _get_anchor(self, year: int) -> pd.Timestamp:
         """Method to generate an anchor timestamp for your specific calendar.
 
@@ -52,34 +54,46 @@ class BaseCalendar(ABC):
         Returns:
             pd.Timestamp: Timestamp at the end of the anchor_years interval 0.
         """
-        return pd.Timestamp()
+        return pd.to_datetime(
+            f"{year}-" + self._anchor, format="%Y-" + self._anchor_fmt
+        )
 
-    def set_max_lag(self, max_lag: int, allow_overlap: bool = False) -> None:
-        """Set the maximum lag of a calendar.
-
-        Sets the maximum number of lag periods after the target period. If `0`,
-        the maximum lag will be determined by how many fit in each anchor year.
-        If a maximum lag is provided, the intervals can either only cover part
-        of the year, or extend over multiple years. In case of a large max_lag
-        number where the intervals extend over multiple years, anchor years will
-        be skipped to avoid overlapping intervals. To allow overlapping
-        intervals, use the `allow_overlap` kwarg.
+    def _parse_anchor(self, anchor_str: str) -> Tuple[str, str]:
+        """Parses the user-input anchor.
 
         Args:
-            max_lag: Maximum number of lag periods after the target period.
-            allow_overlap: Allows intervals to overlap between anchor years, if the
-                max_lag is set to a high enough number that intervals extend over
-                multiple years. `False` by default, to avoid train/test information
-                leakage.
-        """
-        if (max_lag < 0) or (max_lag % 1 > 0):
-            raise ValueError(
-                "Max lag should be an integer with a value of 0 or greater"
-                f", not {max_lag} of type {type(max_lag)}."
-            )
+            anchor_str: Anchor string in the right formatting.
 
-        self._max_lag = max_lag
-        self._allow_overlap = allow_overlap
+        Returns:
+            Datetime formatter to parse the anchor into a date.
+        """
+        if re.fullmatch("\\d{1,2}-\\d{1,2}", anchor_str):
+            fmt = "%m-%d"
+        elif re.fullmatch("\\d{1,2}", anchor_str):
+            fmt = "%m"
+        elif re.fullmatch("W\\d{1,2}-\\d", anchor_str):
+            fmt = "W%W-%w"
+        elif re.fullmatch("W\\d{1,2}", anchor_str):
+            fmt = "W%W-%w"
+            anchor_str += "-1"
+        elif anchor_str.lower() in utils.get_month_names():
+            anchor_str = str(utils.get_month_names()[anchor_str.lower()])
+            fmt = "%m"
+        else:
+            raise ValueError(f"Anchor input '{anchor_str}' does not match expected format")
+        return anchor_str, fmt    
+
+    def _append(self, period_block):
+        """Append target/precursor periods to the calendar."""
+        if period_block.target:
+            self._targets.append(period_block)
+            # count length
+            self._total_length_target += period_block.length + period_block.gap
+
+        else:
+            self._precursors.append(period_block)
+            # count length
+            self._total_length_precursor += period_block.length + period_block.gap
 
     def _map_year(self, year: int) -> pd.Series:
         """Internal routine to return a concrete IntervalIndex for the given year.
@@ -98,28 +112,42 @@ class BaseCalendar(ABC):
             Pandas Series filled with Intervals of the calendar's frequency, counting
             backwards from the calendar's achor.
         """
-        year_intervals = pd.interval_range(
-            end=self._get_anchor(year),
-            periods=self._get_nintervals(),
-            freq=self.freq,
-        )
+        intervals_target = self._concatenate_periods(year, self._targets, True)
+        intervals_precursor = self._concatenate_periods(year, self._precursors, False)
 
-        year_intervals = pd.Series(year_intervals[::-1], name=str(year))
+        self.n_targets = len(intervals_target)
+        year_intervals = intervals_precursor[::-1] + intervals_target
+
+        # turn the list of intervals into pandas series
+        year_intervals = pd.Series(year_intervals[::-1], name=year)
         year_intervals.index.name = "i_interval"
         return year_intervals
 
-    def _get_nintervals(self) -> int:
-        """Calculates the number of intervals that should be generated by _map year.
+    def _concatenate_periods(self, year, list_periods, is_target):
+        # generate intervals based on the building blocks
+        intervals = []
+        if is_target:
+            # build from left to right
+            left_date = self._get_anchor(year)
+            # loop through all the building blocks to
+            for block in list_periods:
+                left_date += pd.Timedelta(block.gap, unit="D")
+                right_date = left_date + pd.Timedelta(block.length, unit="D")
+                intervals.append(pd.Interval(left_date, right_date, closed="left"))
+                # update left date
+                left_date = right_date
+        else:
+            # build from right to left
+            right_date = self._get_anchor(year)
+            # loop through all the building blocks to
+            for block in list_periods:
+                right_date -= pd.Timedelta(block.gap, unit="D")
+                left_date = right_date - pd.Timedelta(block.length, unit="D")
+                intervals.append(pd.Interval(left_date, right_date, closed="left"))
+                # update right date
+                right_date = left_date
 
-        Returns:
-            int: Number of intervals for one anchor year.
-        """
-        periods_per_year = pd.Timedelta("365days") / pd.to_timedelta(self.freq)
-        return (
-            (self._max_lag + self.n_targets)
-            if self._max_lag > 0
-            else int(periods_per_year)
-        )
+        return intervals
 
     def _get_skip_nyears(self) -> int:
         """Determine how many years need to be skipped to avoid overlapping data.
@@ -129,14 +157,11 @@ class BaseCalendar(ABC):
         Returns:
             int: Number of years that need to be skipped.
         """
-        nintervals = self._get_nintervals()
-        periods_per_year = pd.Timedelta("365days") / pd.to_timedelta(self.freq)
+        years = pd.to_timedelta(
+            self._total_length_target + self._total_length_precursor
+        ) / pd.Timedelta("365days")
 
-        return (
-            0
-            if self._max_lag > 0 and self._allow_overlap
-            else int(np.ceil(nintervals / periods_per_year).astype(int) - 1)
-        )
+        return 0 if self._allow_overlap else int(np.ceil(years).astype(int) - 1)
 
     def map_years(self, start: int, end: int):
         """Adds a start and end year mapping to the calendar.
@@ -213,18 +238,30 @@ class BaseCalendar(ABC):
 
         return self
 
-    def _label_targets(self, intervals: pd.DataFrame) -> pd.DataFrame:
+    def _rename_intervals(self, intervals: pd.DataFrame) -> pd.DataFrame:
         """Adds target labels to the header row of the intervals.
 
         Args:
             intervals (pd.Dataframe): Dataframe with intervals.
 
         Returns:
-            pd.Dataframe: Dataframe with target periods labelled.
+            pd.Dataframe: Dataframe with target periods labelled, sorted by i_interval value.
         """
-        return intervals.rename(
-            columns={i: f"(target) {i}" for i in range(self.n_targets)}
+
+        # rename precursors
+        intervals = intervals.rename(
+            columns={
+                i: self.n_targets - i - 1
+                for i in range(self.n_targets, len(intervals.columns))
+            }
         )
+
+        # rename targets
+        intervals = intervals.rename(
+            columns={i: self.n_targets - i for i in range(self.n_targets)}
+        )
+
+        return intervals.sort_index(axis=1)
 
     def get_intervals(self) -> pd.DataFrame:
         """Method to retrieve updated intervals from the Calendar object."""
@@ -242,8 +279,10 @@ class BaseCalendar(ABC):
 
         intervals = pd.concat([self._map_year(year) for year in year_range], axis=1).T
 
+        intervals = self._rename_intervals(intervals)
+
         intervals.index.name = "anchor_year"
-        return intervals
+        return intervals.sort_index(axis=0, ascending=False)
 
     def show(self) -> pd.DataFrame:
         """Displays the intervals the Calendar will generate for the current setup.
@@ -252,7 +291,7 @@ class BaseCalendar(ABC):
             pd.Dataframe: Dataframe containing the calendar intervals, with the target
                 periods labelled.
         """
-        return self._label_targets(self.get_intervals())
+        return self.get_intervals()
 
     def __repr__(self) -> str:
         """String representation of the Calendar."""
@@ -310,3 +349,32 @@ class BaseCalendar(ABC):
     def flat(self) -> pd.DataFrame:
         """Returns the flattened intervals."""
         return self.get_intervals().stack()  # type: ignore
+
+
+class Period(ABC):
+    """Basic construction element of calendar for defining target period."""
+
+    def __init__(self, length: int, gap: int = 0, target: bool = False) -> None:
+        self.length = length
+        self.gap = gap
+        self.target = target
+        # TO DO: support lead_time
+        # self.lead_time = lead_time
+
+
+class TargetPeriod(Period):
+    """Instantiate a build block as target period."""
+
+    def __init__(self, length: int, gap: int = 0) -> None:
+        self.length = length
+        self.gap = gap
+        self.target = True
+
+
+class PrecursorPeriod(Period):
+    """Instantiate a build block as precursor period."""
+
+    def __init__(self, length: int, gap: int = 0) -> None:
+        self.length = length
+        self.gap = gap
+        self.target = False
