+++ s2spy/time.py
@@ -49,14 +49,13 @@ Example:
 """
 import calendar as pycalendar
 import re
-from abc import ABC
 from typing import Tuple
-from typing import Union
 import numpy as np
 import pandas as pd
 import xarray as xr
-from . import utils
 from ._base_calendar import BaseCalendar
+from ._base_calendar import PrecursorPeriod
+from ._base_calendar import TargetPeriod
 from ._resample import resample  # pylint: disable=unused-import
 
 
@@ -120,6 +119,19 @@ class AdventCalendar(BaseCalendar):
             pd.Timestamp: Timestamp at the end of the anchor_years interval 0.
         """
         return pd.Timestamp(year, self.month, self.day)
+
+    def _get_nintervals(self) -> int:
+        """Calculates the number of intervals that should be generated by _map year.
+
+        Returns:
+            int: Number of intervals for one anchor year.
+        """
+        periods_per_year = pd.Timedelta("365days") / pd.to_timedelta(self.freq)
+        return (
+            (self._max_lag + self.n_targets)
+            if self._max_lag > 0
+            else int(periods_per_year)
+        )
 
 
 class MonthlyCalendar(BaseCalendar):
@@ -224,7 +236,7 @@ class MonthlyCalendar(BaseCalendar):
             pd.Dataframe: Dataframe containing the calendar intervals, with the target
                 periods labelled.
         """
-        return self._label_targets(self.get_intervals()).applymap(
+        return self.get_intervals().applymap(
             self._interval_as_month
         )
 
@@ -306,61 +318,38 @@ class WeeklyCalendar(BaseCalendar):
             pd.Dataframe: Dataframe containing the calendar intervals, with the target
                 periods labelled.
         """
-        return self._label_targets(self.get_intervals()).applymap(
+        return self.get_intervals().applymap(
             self._interval_as_weeknr
         )
 
 
-def _parse_anchor(anchor_str: str) -> Tuple[str, str]:
-    """Parses the user-input anchor.
+def set_max_lag(max_lag: int, allow_overlap: bool = False) -> Tuple[int, bool]:
+    """Set the maximum lag of a calendar.
+
+    Sets the maximum number of lag periods after the target period. If `0`,
+    the maximum lag will be determined by how many fit in each anchor year.
+    If a maximum lag is provided, the intervals can either only cover part
+    of the year, or extend over multiple years. In case of a large max_lag
+    number where the intervals extend over multiple years, anchor years will
+    be skipped to avoid overlapping intervals. To allow overlapping
+    intervals, use the `allow_overlap` kwarg.
 
     Args:
-        anchor_str: Anchor string in the right formatting.
-
-    Returns:
-        Datetime formatter to parse the anchor into a date.
+        max_lag: Maximum number of lag periods after the target period.
+        allow_overlap: Allows intervals to overlap between anchor years, if the
+            max_lag is set to a high enough number that intervals extend over
+            multiple years. `False` by default, to avoid train/test information
+            leakage.
     """
-    if re.fullmatch("\\d{1,2}-\\d{1,2}", anchor_str):
-        fmt = "%m-%d"
-    elif re.fullmatch("\\d{1,2}", anchor_str):
-        fmt = "%m"
-    elif re.fullmatch("W\\d{1,2}-\\d", anchor_str):
-        fmt = "W%W-%w"
-    elif re.fullmatch("W\\d{1,2}", anchor_str):
-        fmt = "W%W-%w"
-        anchor_str += "-1"
-    elif anchor_str.lower() in utils.get_month_names():
-        anchor_str = str(utils.get_month_names()[anchor_str.lower()])
-        fmt = "%m"
-    else:
-        raise ValueError(f"Anchor input '{anchor_str}' does not match expected format")
-    return anchor_str, fmt
+    if (max_lag < 0) or (max_lag % 1 > 0):
+        raise ValueError(
+            "Max lag should be an integer with a value of 0 or greater"
+            f", not {max_lag} of type {type(max_lag)}."
+        )
 
-
-def _rename_intervals(calendar, intervals: pd.DataFrame) -> pd.DataFrame:
-    """Adds target labels to the header row of the intervals.
-
-    Args:
-        intervals (pd.Dataframe): Dataframe with intervals.
-
-    Returns:
-        pd.Dataframe: Dataframe with target periods labelled, sorted by i_interval value.
-    """
-
-    # rename precursors
-    intervals = intervals.rename(
-        columns={
-            i: calendar.n_targets - i - 1
-            for i in range(calendar.n_targets, len(intervals.columns))
-        }
-    )
-
-    # rename targets
-    intervals = intervals.rename(
-        columns={i: calendar.n_targets - i for i in range(calendar.n_targets)}
-    )
-
-    return intervals.sort_index(axis=1)
+    # self._max_lag = max_lag
+    # self._allow_overlap = allow_overlap
+    return max_lag, allow_overlap
 
 
 class CustomCalendar(BaseCalendar):
@@ -389,7 +378,7 @@ class CustomCalendar(BaseCalendar):
             n_targets (int): Number of targets that inferred from the appended
             `TargetPeriod` blocks.
         """
-        self._anchor, self._anchor_fmt = _parse_anchor(anchor)
+        self._anchor, self._anchor_fmt = self._parse_anchor(anchor)
         self._targets: list[TargetPeriod] = []
         self._precursors: list[PrecursorPeriod] = []
         self._total_length_target = 0
@@ -398,179 +387,12 @@ class CustomCalendar(BaseCalendar):
 
         self._allow_overlap: bool = False
 
-    def _get_anchor(self, year: int) -> pd.Timestamp:
-        """Generates a timestamp for the end of interval 0 in year.
-
-        Args:
-            year (int): anchor year for which the anchor timestamp should be generated
-
-        Returns:
-            pd.Timestamp: Timestamp at the end of the anchor_years interval 0.
-        """
-        return pd.to_datetime(
-            f"{year}-" + self._anchor, format="%Y-" + self._anchor_fmt
-        )
-
     def append(self, period_block):
-        """Append target/precursor periods to the calendar."""
-        if period_block.target:
-            self._targets.append(period_block)
-            # count length
-            self._total_length_target += period_block.length + period_block.gap
-
-        else:
-            self._precursors.append(period_block)
-            # count length
-            self._total_length_precursor += period_block.length + period_block.gap
-
-    def _map_year(self, year: int):
-        """Replace old map_year function"""
-        intervals_target = self._concatenate_periods(year, self._targets, True)
-        intervals_precursor = self._concatenate_periods(year, self._precursors, False)
-
-        self.n_targets = len(intervals_target)
-        year_intervals = intervals_precursor[::-1] + intervals_target
-
-        # turn the list of intervals into pandas series
-        year_intervals = pd.Series(year_intervals[::-1], name=year)
-        year_intervals.index.name = "i_interval"
-        return year_intervals
-
-    def _concatenate_periods(self, year, list_periods, is_target):
-        # generate intervals based on the building blocks
-        intervals = []
-        if is_target:
-            # build from left to right
-            left_date = self._get_anchor(year)
-            # loop through all the building blocks to
-            for block in list_periods:
-                left_date += pd.Timedelta(block.gap, unit="D")
-                right_date = left_date + pd.Timedelta(block.length, unit="D")
-                intervals.append(pd.Interval(left_date, right_date, closed="left"))
-                # update left date
-                left_date = right_date
-        else:
-            # build from right to left
-            right_date = self._get_anchor(year)
-            # loop through all the building blocks to
-            for block in list_periods:
-                right_date -= pd.Timedelta(block.gap, unit="D")
-                left_date = right_date - pd.Timedelta(block.length, unit="D")
-                intervals.append(pd.Interval(left_date, right_date, closed="left"))
-                # update right date
-                right_date = left_date
-
-        return intervals
-
-    def map_years(self, start: int, end: int):
-        """Adds a start and end year mapping to the calendar.
-
-        If the start and end years are the same, the intervals for only that single
-        year are returned by calendar.get_intervals().
-
-        Args:
-            start: The first year for which the calendar will be realized
-            end: The last year for which the calendar will be realized
-
-        Returns:
-            The calendar mapped to the input start and end year.
+        """Append target/precursor periods to the calendar.
+        
+        Wrapper function of append method.
         """
-        if start > end:
-            raise ValueError("The start year cannot be greater than the end year")
-
-        self._first_year = start
-        self._last_year = end
-        self._mapping = "years"
-        return self
-
-    def map_to_data(
-        self,
-        input_data: Union[pd.Series, pd.DataFrame, xr.Dataset, xr.DataArray],
-    ):
-        """Map the calendar to input data period.
-
-        Stores the first and last intervals of the input data to the calendar, so that
-        the intervals can cover the data to the greatest extent.
-
-        Args:
-            input_data: Input data for datetime mapping. Its index must be either
-                pandas.DatetimeIndex, or an xarray `time` coordinate with datetime
-                data.
-
-        Returns:
-            The calendar mapped to the input data period.
-        """
-        utils.check_timeseries(input_data)
-
-        # check the datetime order of input data
-        if isinstance(input_data, PandasData):
-            self._first_timestamp = input_data.index.min()
-            self._last_timestamp = input_data.index.max()
-        else:
-            self._first_timestamp = pd.Timestamp(input_data.time.min().values)
-            self._last_timestamp = pd.Timestamp(input_data.time.max().values)
-
-        self._mapping = "data"
-
-        return self
-
-    def _set_year_range_from_timestamps(self):
-        min_year = self._first_timestamp.year
-        max_year = self._last_timestamp.year
-
-        # ensure that the input data could always cover the advent calendar
-        # last date check
-        if self._map_year(max_year).iloc[0].right > self._last_timestamp:
-            max_year -= 1
-        # first date check
-        while self._map_year(min_year).iloc[-1].right <= self._first_timestamp:
-            min_year += 1
-
-        # map year(s) and generate year realized advent calendar
-        if max_year >= min_year:
-            self._first_year = min_year
-            self._last_year = max_year
-        else:
-            raise ValueError(
-                "The input data could not cover the target advent calendar."
-            )
-
-        return self
-
-    def get_intervals(self) -> pd.DataFrame:
-        """Method to retrieve updated intervals from the Calendar object."""
-        if self._mapping is None:
-            raise ValueError(
-                "Cannot retrieve intervals without map_years or "
-                "map_to_data having configured the calendar."
-            )
-        if self._mapping == "data":
-            self._set_year_range_from_timestamps()
-
-        year_range = range(
-            self._last_year, self._first_year - 1, -(self._get_skip_nyears() + 1)
-        )
-
-        intervals = pd.concat([self._map_year(year) for year in year_range], axis=1).T
-
-        intervals = _rename_intervals(self, intervals)
-
-        intervals.index.name = "anchor_year"
-        return intervals.sort_index(axis=0, ascending=False)
-
-    def _get_skip_nyears(self) -> int:
-        """Determine how many years need to be skipped to avoid overlapping data.
-
-        Required to prevent information leakage between anchor years.
-
-        Returns:
-            int: Number of years that need to be skipped.
-        """
-        years = pd.to_timedelta(
-            self._total_length_target + self._total_length_precursor
-        ) / pd.Timedelta("365days")
-
-        return 0 if self._allow_overlap else int(np.ceil(years).astype(int) - 1)
+        self._append(period_block)
 
     def show(self) -> pd.DataFrame:
         """Displays the intervals the Calendar will generate for the current setup.
@@ -580,32 +402,3 @@ class CustomCalendar(BaseCalendar):
                 periods labelled.
         """
         return self.get_intervals()
-
-
-class Period(ABC):
-    """Basic construction element of calendar for defining target period."""
-
-    def __init__(self, length: int, gap: int = 0, target: bool = False) -> None:
-        self.length = length
-        self.gap = gap
-        self.target = target
-        # TO DO: support lead_time
-        # self.lead_time = lead_time
-
-
-class TargetPeriod(Period):
-    """Instantiate a build block as target period."""
-
-    def __init__(self, length: int, gap: int = 0) -> None:
-        self.length = length
-        self.gap = gap
-        self.target = True
-
-
-class PrecursorPeriod(Period):
-    """Instantiate a build block as precursor period."""
-
-    def __init__(self, length: int, gap: int = 0) -> None:
-        self.length = length
-        self.gap = gap
-        self.target = False
